requirements.txt
aiohttp==3.12.14
apscheduler
twikit==3.11.0
beautifulsoup4==4.13.4
discord.py==2.5.2
mcrcon==0.7.0
mcstatus==12.0.2
psutil==7.0.0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
